all: move config package to new module
This change looks much bigger than it is.
