+++ config/go.mod
@@ -1,0 +1,7 @@
+module github.com/quay/clair/config
+
+go 1.16
+
+require github.com/google/go-cmp v0.5.6
+
+require golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
