+++ notifier/postgres/setdelivered.go
@@ -8,6 +8,7 @@ import (
 	"github.com/jackc/pgx/v4/pgxpool"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promauto"
+
 	clairerror "github.com/quay/clair/v4/clair-error"
 )
