+++ notifier/stomp/directdeliverer.go
@@ -8,9 +8,9 @@ import (
 
 	gostomp "github.com/go-stomp/stomp"
 	"github.com/google/uuid"
+	"github.com/quay/clair/config"
 
 	clairerror "github.com/quay/clair/v4/clair-error"
-	"github.com/quay/clair/v4/config"
 	"github.com/quay/clair/v4/notifier"
 )
