+++ notifier/stomp/directdeliverer_integration_test.go
@@ -10,12 +10,12 @@ import (
 
 	"github.com/go-stomp/stomp"
 	"github.com/google/uuid"
+	"github.com/quay/clair/config"
 	"github.com/quay/claircore"
 	"github.com/quay/claircore/test/integration"
 	"github.com/quay/zlog"
 	"golang.org/x/sync/errgroup"
 
-	"github.com/quay/clair/v4/config"
 	"github.com/quay/clair/v4/notifier"
 )
